+++ Packages/com.dss.core-utils/Runtime/ApplicationUtilities/OpenURL.cs
@@ -4,6 +4,7 @@ using UnityEngine.EventSystems;
 namespace DSS.CoreUtils.ApplicationUtilities
 {
     // @brief Exposes a method to open a URL in the users web browser.
+    [AddComponentMenu("DSS/Application Utilities/Open URL")]	
     public class OpenURL : MonoBehaviour
     {
         [SerializeField] string url = "https://www.google.com/";