+++ Packages/com.dss.core-utils/Runtime/ApplicationUtilities/Quiter.cs
@@ -3,6 +3,7 @@ using UnityEngine;
 namespace DSS.CoreUtils.ApplicationUtilities
 {
     // @brief Exposes a method to quit the application.
+    [AddComponentMenu("DSS/Application Utilities/Quiter")]	
     public class Quiter : MonoBehaviour
     {
         public void Quit()