+++ Packages/com.dss.core-utils/Runtime/ApplicationUtilities/SetTargetFrameRate.cs
@@ -3,6 +3,7 @@ using UnityEngine;
 namespace DSS.CoreUtils.ApplicationUtilities
 {
     // @brief Sets the Application's target framerate when the scene first loads.
+    [AddComponentMenu("DSS/Application Utilities/Set Target Frame Rate")]	
     public class SetTargetFrameRate : MonoBehaviour
     {
         // @brief The target framerate.