+++ Packages/com.dss.core-utils/Runtime/Events/OnKeyCodeCombination.cs
@@ -4,6 +4,7 @@ using UnityEngine.Events;
 namespace DSS.CoreUtils.Events
 {
     // @brief A helper class that exposes an event to the editor that runs on scene load.
+    [AddComponentMenu("DSS/Events/On KeyCode Combination")]	
     public class OnKeyCodeCombination : MonoBehaviour
     {
         [SerializeField] UnityEvent onKeyCodeCombination = default;